Added documentation and Remove SonarLint warnings

+++ src/main/java/cz/muni/fi/pb162/project/BoardNotation.java
@@ -21,7 +21,7 @@ public final class BoardNotation {
     }
 
     private static int getNumberFromLetter(char letter) {
-        return (int) letter - UNICODE_VALUE_FOR_A;
+        return letter - UNICODE_VALUE_FOR_A;
     }
 
     /**
